template/backend/requirements.txt
# Core agent framework
# agent-framework[all] pulls in orchestrations, azure-ai-search, a2a, etc.
agent-framework[all]>=1.0.0
agent-framework-foundry>=1.0.0

# Azure SDK
azure-identity>=1.21.0
azure-cosmos>=4.7.0
azure-keyvault-secrets>=4.9.0
azure-monitor-opentelemetry>=1.6.0

# Azure AI Foundry SDK
azure-ai-projects>=1.0.0

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
websockets>=13.1

# Authentication
python-jose[cryptography]>=3.3.0
httpx>=0.28.0

# Observability
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-otlp-proto-grpc>=1.30.0
opentelemetry-instrumentation-fastapi>=0.51b0
opentelemetry-semantic-conventions-ai>=0.5.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.8.0

# Evaluation
azure-ai-evaluation>=1.0.0
